requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email] 
python-multipart
pytest
pytest-asyncio
httpx
bcrypt
